refactor: separate navigation and general volume views
